Fixed some of the integer widths, fleshed out fmt, added more functions to runtime.

+++ types.h
@@ -3,13 +3,13 @@
 
 typedef signed char i8;
 typedef short       i16;
-typedef long        i32;
-typedef long long   i64;
+typedef int         i32;
+typedef long        i64;
 
-typedef unsigned char      u8;
-typedef unsigned short     u16;
-typedef unsigned long      u32;
-typedef unsigned long long u64;
+typedef unsigned char  u8;
+typedef unsigned short u16;
+typedef unsigned int   u32;
+typedef unsigned long  u64;
 
 typedef u64 usize;
 typedef i64 isize;
